api/servers: initial refactor (mostly just a mv)
